+++ mrb_doc/core/window.rb
@@ -289,6 +289,15 @@ def get_monitor_height(monitor)
   1080
 end
 
+# Gets the refresh rate of the specified monitor
+# @param monitor [Integer]
+# @return [Integer]
+def get_monitor_refresh_rate(monitor)
+  # mrb_get_monitor_refresh_rate
+  # src/mruby_integration/core/window.cpp
+  60
+end
+
 # Gets the position of the window
 # @return [Vector2]
 def get_window_position()