+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Taylor
 
+## Unreleased
+
 ## v0.0.6
 
 - Add draw_triangle_lines